requirements.txt
python-dotenv
requests
langchain
langchain-groq
